Disable tsconfig options that cause vscode type checking to fail
For some reason, `--noUnusedLocals` will cause the vscode type checking
in tsc 2.7.1 to fail in the soon-to-be-committed file,
`src/ql/query.ts`.  Specifically, errors will stop being reported, and
when they are reported, they often refer to errors that were recently
fixed. For example, if you press `return` several times, the errors can
be reported from inside comments.

It's not clear why, but we were able to verify that flipping this off
will specifically cause type checking to work.